Update newView, use newView and disable buttons
Update the method newView in the class ChangeView so that it uses the path argument, previously the path argument wasn't used. Use the method newView in the methods actionSave and actionCancel in the class ContactController, previously the code to changeView was pasted into these methods instead of the method newView being used. In the method actionSearch in the class ListOfContactsController, disable the Update and Delete buttons after a search has been made.